specs: update predeploy spec
It fell out of date, so this commit updates the spec
to be more modern. It also removes the inlined solidity
interfaces for the contracts and instead adds a link
to the contract itself. This will help to prevent churn
with maintaining a copy of the interface inlined in
the spec. Some language is updated when describing each
predeploy. There is a tradeoff in having the contracts
themselves explain each function vs the spec explaining
each function.